api/requirements.txt
# Framework et API
flask==2.2.3
flask-cors==3.0.10
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.9.2,<2.10.0
gunicorn>=21.2.0

# Requêtes et Authentification
requests==2.28.2
passlib==1.7.4
python-jose==3.3.0
python-multipart>=0.0.7
aiohttp>=3.8.5

# Base de données et ORM
sqlalchemy==2.0.39
alembic>=1.12.0
redis==4.6.0

# IA et LLM
ollama==0.1.2
# Désactivation temporaire de oumi pour résoudre le conflit de dépendances
# oumi>=0.1.3,<0.1.6
transformers>=4.35.0
tokenizers>=0.14.0
sentence-transformers>=2.2.2
datasets>=2.14.0
scikit-learn>=1.3.0
torch>=2.0.0
gensim>=4.3.1
crewai>=0.19.0

# Web Scraping et NLP
scrapy==2.10.0
beautifulsoup4>=4.12.2
selectolax>=0.3.16
lxml>=4.9.3
playwright>=1.40.0
dateutil>=2.8.2
pandas>=2.0.0
openai>=1.3.0  # Pour compatibilité avec certaines bibliothèques

# Traitement de texte et IA
spacy>=3.7.0
nltk>=3.8.1
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.18

# Indexation et Recherche
elasticsearch==8.9.0
faiss-cpu>=1.7.4  # Pour la recherche vectorielle rapide

# Utilitaires
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.66.0
torchdata==0.7.1
regex>=2023.8.8
backoff>=2.2.1
